makes Fortran's type integer(4) represented in 4 bytes (corresponding to C's type int).

+++ XcodeML-Exc-Tools/src/exc/xmpF/XMP.java
@@ -135,7 +135,7 @@ public class XMP {
 	  t = BasicType.SHORT;
 	  break;
 	case 4:
-	  t = BasicType.LONG;
+	  t = BasicType.INT;
 	  break;
 	case 8:
 	  t = BasicType.LONGLONG;
